Refactor into dedicated files and companion objects
